feat: Deploy ManualToken contract and update frontend configuration
- Added deployment details for ManualToken contract in run-latest.json.
- Updated frontend configuration to use the new deployed contract address.
- Modified DeployManualToken script to ensure proper handling of the deployer private key.

+++ script/DeployManualToken.s.sol
@@ -6,7 +6,7 @@ import {ManualToken} from "../src/ManualToken.sol";
 
 contract DeployManualToken is Script {
     function run() external returns (ManualToken) {
-        uint256 deployerPrivateKey = vm.envUint("PRIVATE_KEY");
+        uint256 deployerPrivateKey = uint256(vm.envUint("PRIVATE_KEY"));
         vm.startBroadcast(deployerPrivateKey);
 
         ManualToken token = new ManualToken("ManualToken", "MTK", 1000000);
